+++ kfs-4/include/paging.h
@@ -26,15 +26,15 @@
 
 #define PG_RESERVED_ENTRY    (PG_RESERVED | PG_PS | PG_RDWR)
 
-#define addr_from_pte(pte)          ((((uint32_t)(pte) & 0x003FF000) << 10) | (((uint32_t)(pte) & 0x00000FFF) << 10))
-#define pte_from_addr(addr)         (0xFFC00000 | (((uint32_t)(addr) & 0xFFC00000) >> 10) | (((uint32_t)(addr) & 0x003FF000) >> 10))
-#define addr_from_pde(dir)          (((uint32_t)(dir) & 0x00000FFF) << 20)
-#define pde_from_addr(addr)         (0xFFFFF000 | (((uint32_t)(addr) & 0xFFC00000) >> 20))
-#define addr_erase_offset(addr)     ((uint32_t)(addr) & 0xFFFFF000)
-#define addr_get_offset(addr)       ((uint32_t)(addr) & 0x00000FFF)
-#define k_addr_erase_offset(addr)   ((uint32_t)(addr) & 0xFFC00000)
-#define k_addr_get_offset(addr)     ((uint32_t)(addr) & 0x003FFFFF)
-#define is_kernel_space(addr)       ((uint32_t)(addr) >= K_VSPACE_START)
+#define addr_from_pte(pte)          ((((uintptr_t)(pte) & 0x003FF000) << 10) | (((uintptr_t)(pte) & 0x00000FFF) << 10))
+#define pte_from_addr(addr)         (0xFFC00000 | (((uintptr_t)(addr) & 0xFFC00000) >> 10) | (((uintptr_t)(addr) & 0x003FF000) >> 10))
+#define addr_from_pde(dir)          (((uintptr_t)(dir) & 0x00000FFF) << 20)
+#define pde_from_addr(addr)         (0xFFFFF000 | (((uintptr_t)(addr) & 0xFFC00000) >> 20))
+#define addr_erase_offset(addr)     ((uintptr_t)(addr) & 0xFFFFF000)
+#define addr_get_offset(addr)       ((uintptr_t)(addr) & 0x00000FFF)
+#define k_addr_erase_offset(addr)   ((uintptr_t)(addr) & 0xFFC00000)
+#define k_addr_get_offset(addr)     ((uintptr_t)(addr) & 0x003FFFFF)
+#define is_kernel_space(addr)       ((uintptr_t)(addr) >= K_VSPACE_START)
 
 static inline void tlb_flush_all(void) 
 {
@@ -44,7 +44,7 @@ static inline void tlb_flush_all(void)
     __asm__ volatile ("mov %0, %%cr3" :: "r" (cr3));
 }
 
-static inline void tlb_flush(uint32_t addr) 
+static inline void tlb_flush(uintptr_t addr) 
 {
     __asm__ volatile ("invlpg (%0)" :: "r" (addr));
 }
